rename: storybook 설정변경 및 assets 경로변경
