+++ app/_asset/fonts/PretendardVariable.woff2
