+++ app/fonts.ts
@@ -1,7 +1,7 @@
 import localFont from 'next/font/local';
 
 export const Pretendard = localFont({
-  src: '../public/fonts/PretendardVariable.woff2',
+  src: './_asset/fonts/PretendardVariable.woff2',
   display: 'swap',
   variable: '--font-pretendard', // css 변수로 등록하기 위함
 });
