+++ components/core/HelperText/HelperText.tsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const HelperText = () => {
+  return (
+    <div></div>
+  )
+}
+
+export default HelperText
