+++ components/core/Label/Label.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Label = () => {
+  return <div></div>;
+};
+
+export default Label;
